backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
redis==5.2.0
psycopg2-binary==2.9.10
